Modulo 2/apiperu-etl/requirements.txt
aiosqlite==0.20.0
alembic==1.13.2
annotated-types==0.7.0
anyio==3.7.1
apprise==1.8.1
asgi-lifespan==2.1.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
coolname==2.2.0
croniter==2.0.7
cryptography==43.0.0
dateparser==1.2.0
dnspython==2.6.1
docker==7.1.0
email_validator==2.2.0
fsspec==2024.6.1
google-auth==2.32.0
graphviz==0.20.3
greenlet==3.0.3
griffe==0.47.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
humanize==4.10.0
hyperframe==6.0.1
idna==3.7
importlib_resources==6.1.3
itsdangerous==2.2.0
Jinja2==3.1.4
jinja2-humanize-extension==0.4.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kubernetes==29.0.0
Mako==1.3.5
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mysql-connector-python==9.0.0
oauthlib==3.2.2
orjson==3.10.6
packaging==24.1
pathspec==0.12.1
pendulum==2.1.2
prefect==2.19.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-multipart==0.0.9
python-slugify==8.0.4
pytz==2024.1
pytzdata==2020.1
PyYAML==6.0.1
readchar==4.1.0
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rich==13.7.1
rpds-py==0.19.1
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
text-unidecode==1.3
toml==0.10.2
typer==0.12.3
typing_extensions==4.12.2
tzlocal==5.2
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.3
websocket-client==1.8.0
websockets==12.0
